Add files via upload
This file, rpm_gradle.zip is an export from an eclipse project.  It is a project we use except that everything has been stripped out of it.  There is no code.  The build.gradle is pretty much intact except it doesn't point to my corporate Maven repo, but to JCenter and the dependency on gradle.ospackage has been reset to a publicly available version.  Nonetheless, it still exhibits the same problem:  As it is now both the Java build path (.classpath) and the compiler level are set to jdk-1.8.  Simply running Gradle ->Refresh Gradle Project changes everything to 1.7